requirements.txt
#requirements for python-3.3.0
Django==1.5
distribute==0.6.35
factory-boy==2.0.0-dev
